use pipewire+pipewire-pulse as default sound server
Based on discussion from https://bugs.freebsd.org/290403, the
decision to switch to Pipewire as the default sound server will
require more work, design, and approval.

In the meantime, since it is required for screensharing on
wayland, use the built-in pipewire-pulse daemon to provide
Pulseaudio compatibility. This patch will disable the XDG
autostart entry for pulseaudio by setting Hidden=true in
its user entry.